# Devirtualize methods for -State combinator.