--- a/dbms/src/AggregateFunctions/AggregateFunctionState.cpp
+++ b/dbms/src/AggregateFunctions/AggregateFunctionState.cpp
@@ -1,4 +1,3 @@
-#include <AggregateFunctions/AggregateFunctionState.h>
 #include <AggregateFunctions/AggregateFunctionMerge.h>
 #include <AggregateFunctions/AggregateFunctionCombinatorFactory.h>
 #include <DataTypes/DataTypeAggregateFunction.h>
@@ -6,12 +5,6 @@
 
 namespace DB
 {
-
-namespace ErrorCodes
-{
-    extern const int NUMBER_OF_ARGUMENTS_DOESNT_MATCH;
-    extern const int BAD_ARGUMENTS;
-}
 
 class AggregateFunctionCombinatorState final : public IAggregateFunctionCombinator
 {
@@ -24,47 +17,17 @@ public:
     }
 
     AggregateFunctionPtr transformAggregateFunction(
-        const AggregateFunctionPtr & nested_function, const DataTypes & arguments, const Array & params) const override
+        const AggregateFunctionPtr & nested_function, const DataTypes &, const Array &) const override
     {
-        return std::make_shared<AggregateFunctionState>(nested_function, arguments, params);
+        auto ret = nested_function->getInstance();
+        ret->setState(true);
+        return ret;
     }
 };
 
 void registerAggregateFunctionCombinatorState(AggregateFunctionCombinatorFactory & factory)
 {
     factory.registerCombinator(std::make_shared<AggregateFunctionCombinatorState>());
-}
-
-
-DataTypePtr AggregateFunctionState::getReturnType() const
-{
-    auto ptr = std::make_shared<DataTypeAggregateFunction>(nested_func, arguments, params);
-
-    /// Special case: it is -MergeState combinator.
-    /// We must return AggregateFunction(agg, ...) instead of AggregateFunction(aggMerge, ...)
-    if (typeid_cast<const AggregateFunctionMerge *>(ptr->getFunction().get()))
-    {
-        if (arguments.size() != 1)
-            throw Exception("Combinator -MergeState expects only one argument", ErrorCodes::BAD_ARGUMENTS);
-
-        if (!typeid_cast<const DataTypeAggregateFunction *>(arguments[0].get()))
-            throw Exception("Combinator -MergeState expects argument with AggregateFunction type", ErrorCodes::BAD_ARGUMENTS);
-
-        return arguments[0];
-    }
-    if (arguments.size() > 0)
-    {
-        DataTypePtr argument_type_ptr = arguments[0];
-        WhichDataType which(*argument_type_ptr);
-        if (which.idx == TypeIndex::AggregateFunction)
-        {
-            if (arguments.size() != 1)
-                throw Exception("Nested aggregation expects only one argument", ErrorCodes::BAD_ARGUMENTS);
-            return arguments[0];
-        }
-    }
-
-    return ptr;
 }
 
 }