@filimonov 2019-10-28
Some tests failed.

BTW - interesting how does it compare now with SimpleAggregateFunction ( https://github.com/ClickHouse/ClickHouse/pull/4629 )