@amosbird 2019-10-29
> BTW - interesting how does it compare now with SimpleAggregateFunction

tbh, I fail to see the correlation.