@filimonov 2019-10-29
SimpleAggregationFunction were introduced due to the big overhead of states processing of usual AggregationFunction. I'm wondering if were is a sense of using SimpleAggregationFunction after your change.